+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [WIP]
+### Added
+- Support Redirection and History
+
 ## [0.3.0] - 2017-11-27
 ### Added
 - Support POST Multiple Multipart-Encoded Files #8